+++ mail/mail-uniffi/CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
   - Added `Mailbox::all_mail()` constructor.
+  - Added `Draft::save()`
+  - Added `Draft::set_subject()`
+  - Added `Draft::set_body()`
+  - Added `Draft::set_to_recipients()`
+  - Added `Draft::set_cc_recipients()`
+  - Added `Draft::set_bcc_recipients()`
 
 ### Changed
 
@@ -20,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     `all_available_bottom_bar_actions_for_conversations()` now contains
     `local_id` for `Labels`.
   - Remove damping on account and session watchers
+  - Drafts only create data on the first call to `Draft::save()`
 
 
 ### Fixed